File optimize tests.
